+++ custom_addons/ak_library_management/wizard/borrow_transaction_history_wizard.py
@@ -18,3 +18,14 @@ class BorrowTransactionHistory(models.TransientModel):
         """
         rec_id = self.env.context.get('active_id')
         self.env["borrow.transaction.history"].browse(rec_id).unlink()
+
+    def action_continue(self):
+        """
+        if whenever click continue button then decrease on hand quantity by one.
+        """
+        books = self.env.context.get('book_ids')
+        for rec in self.env["product.template"].browse(books).filtered(lambda book:
+                                                                       book.qty_available):
+            loc = self.env['stock.quant'].search([('product_tmpl_id.id', '=', rec.id)], limit=1)
+            self.env['stock.quant']._update_available_quantity(loc.product_id, loc.location_id,
+                                                               quantity=-1)