+++ custom_addons/ak_library_management/changelog.md
@@ -26,3 +26,4 @@
 
 - change the domain of decrease on hand quantity functionality
 - use the filtered method for code optimization
+- change in continue button functionality